Move some test utils to test/util
