+++ src/layman/layer/geoserver/geoserver_test.py
@@ -2,7 +2,6 @@ from test import process_client, util as test_util, data as test_data
 import pytest
 
 from layman import app, settings
-from layman.common import bbox as bbox_util
 from layman.common.prime_db_schema import publications
 from layman.http import LaymanError
 from . import wfs, wms
@@ -31,21 +30,6 @@ def test_check_user_wms():
     assert exc_info.value.data['workspace_name'] == user
 
 
-def assert_wfs_bbox(workspace, layer, expected_bbox):
-    wfs_layer = f"{workspace}:{layer}"
-    wfs_get_capabilities = wfs.get_wfs_proxy(workspace)
-    wfs_bbox_4326 = wfs_get_capabilities.contents[wfs_layer].boundingBoxWGS84
-    wfs_bbox_3857 = bbox_util.transform(wfs_bbox_4326, 4326, 3857, )
-    test_util.assert_same_bboxes(expected_bbox, wfs_bbox_3857, 0.00001)
-
-
-def assert_wms_bbox(workspace, layer, expected_bbox):
-    wms_get_capabilities = wms.get_wms_proxy(workspace)
-    wms_bboxes = wms_get_capabilities.contents[layer].crs_list
-    wms_bbox_3857 = next(bbox[:4] for bbox in wms_bboxes if bbox[4] == 'EPSG:3857')
-    test_util.assert_same_bboxes(expected_bbox, wms_bbox_3857, 0.00001)
-
-
 @pytest.mark.usefixtures('ensure_layman')
 def test_geoserver_bbox():
     workspace = 'test_geoserver_bbox_workspace'
@@ -65,8 +49,8 @@ def test_geoserver_bbox():
     process_client.publish_workspace_layer(workspace, layer, style_file='sample/style/small_layer.qml')
 
     with app.app_context():
-        assert_wfs_bbox(workspace, layer, expected_bbox_1)
-        assert_wms_bbox(workspace, layer, expected_bbox_1)
+        test_util.assert_wfs_bbox(workspace, layer, expected_bbox_1)
+        test_util.assert_wms_bbox(workspace, layer, expected_bbox_1)
 
         # test WFS
         for bbox, expected_bbox in expected_bboxes:
@@ -74,7 +58,7 @@ def test_geoserver_bbox():
             wfs.delete_layer(workspace, layer)
             geoserver.publish_layer_from_db(workspace, layer, layer, layer, access_rights=None)
             wfs.clear_cache(workspace)
-            assert_wfs_bbox(workspace, layer, expected_bbox)
+            test_util.assert_wfs_bbox(workspace, layer, expected_bbox)
 
         # test WMS
         for bbox, expected_bbox in expected_bboxes:
@@ -88,7 +72,7 @@ def test_geoserver_bbox():
                                             geoserver_workspace=geoserver_workspace,
                                             )
             wms.clear_cache(workspace)
-            assert_wms_bbox(workspace, layer, expected_bbox)
+            test_util.assert_wms_bbox(workspace, layer, expected_bbox)
 
         # test cascade WMS from QGIS
         for bbox, expected_bbox in expected_bboxes:
@@ -102,6 +86,6 @@ def test_geoserver_bbox():
                                               geoserver_workspace=geoserver_workspace,
                                               )
             wms.clear_cache(workspace)
-            assert_wms_bbox(workspace, layer, expected_bbox)
+            test_util.assert_wms_bbox(workspace, layer, expected_bbox)
 
     process_client.delete_workspace_layer(workspace, layer)
